+++ CHANGELOG.md
@@ -5,10 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.4] - 2025-03-28
+
+### Modified
+
+- Refactored all commands to use `SynDataGenYOLO {command}`
 
 ## [0.1.3] - 2025-03-16
 
 ### Added
 
 First release of the package.
+
+- Added the `SynDataGenYOLO` command line interface.